Module 2 Video 1 edited code
In this video I learned uses of dataset, how to create a dataset and how to add examples to it. The main aim of dataset is to set golden examples that we need the llm to get correct. On adding examples I added examples programmatically, manually, by tracing and by AI generation.