Module 2 Video 2 edited code
In this video I learned what an evaluator is, uses of it and different ways to create it. Evaluators basically tests how good the output from a code is by comparing it to our golden example. So it takes input, run output, reference output and spits out score on how good the run output is. I also learned how to define evaluator in local code and llm as judge, custom code in LangSmith UI