Module 2 Video 5 edited code
In this video I learned about comparing experiment result to figure out which output is better and in turn figure out which prompt is better. LLM as judge was used to compare. Pairwise experiment is very useful in cases where if we score the outputs independently and are not able to figure out which is better. This directly compares the outputs head to head so that we can figure out which one is better.