Module 2 Video 6 edited code
In this video I learned about summary evaluators. They are used when we need to calculate a metric over the entire experiment and wont be meaningful when calculating just for one example. We had dealt with getting a score for each example by comparing the run output and reference output but here all these scores of the experiment are needed to calculate the wanted metric. This is where summary evaluators are used.